Decompile menuitem0f0e5ef8

+++ src/include/game/menuitem.h
@@ -5,7 +5,7 @@
 #include "types.h"
 
 u32 menuitem0f0e5d2c(void);
-s32 menuitem0f0e5ef8(s16 arg0, struct menuitem *item);
+s16 menuitem0f0e5ef8(s16 arg0, struct menuitem *item);
 u32 menuitem0f0e6038(void);
 Gfx *menuitemListOverlay(Gfx *gdl, s16 x, s16 y, s16 x2, s16 y2);
 Gfx *menuitemListRender(Gfx *gdl, struct menurendercontext *context);
